Move app and config files to subdirectories
